refactor: 페이지명 수정 (AddAlias -> AddCardAlias
